requirements.txt
pandas
mlflow
numpy
scikit-learn
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
Flask
Flask-Cors
-e .